Moved permalink mapping utilities unto common/full-table/utils/mapPermalinks.ts
